sketch instance types cleanup
